Doctrine\'s contexts moved to separate namespace
